pyproject.toml
[project]
name = "football-news-hub"
version = "0.1.0"
description = "A configurable football news crawler supporting multiple data sources"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.52.0",
    "pydantic>=2.11.0",
    "pyyaml>=6.0.2",
    "typer>=0.15.0",
    "loguru>=0.7.3",
    "fastmcp>=2.3.0",
    "openai>=1.30.0",
    "httpx>=0.27.0",
]

[project.scripts]
football-news = "football_news_hub.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/football_news_hub"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
